fix: changes component
