+++ src/types/interactiveModal/IAction/index.ts
@@ -1,0 +1,10 @@
+interface IAction {
+  id: string;
+  content:
+    | string
+    | ((data: {
+        [key: string]: string | number | boolean | React.ReactNode;
+      }) => React.ReactNode);
+}
+
+export type { IAction };
