+++ src/types/interactiveModal/IField/index.ts
@@ -1,0 +1,7 @@
+interface IField {
+  id: string;
+  labelName: string;
+  type?: string;
+}
+
+export type { IField };
